WIP: refactor & simplify scene loading and colorize
