+++ src/camera.rs
@@ -2,7 +2,7 @@ use crate::{random::random_in_unit_disk, Float, Ray, Vec3, PI};
 use rand::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
 pub struct Camera {
     pub lower_left_corner: Vec3,
     pub horizontal: Vec3,
@@ -72,7 +72,7 @@ impl Camera {
     }
 
     // TODO: fix the mysterious (u,v) vs (s,t) change that came from the tutorial
-    pub fn get_ray(&self, s: Float, t: Float, mut rng: ThreadRng) -> Ray {
+    pub fn get_ray(self, s: Float, t: Float, mut rng: ThreadRng) -> Ray {
         // TODO: add a better defocus blur / depth of field implementation
         let rd: Vec3 = self.lens_radius * random_in_unit_disk(&mut rng);
         let offset: Vec3 = self.u * rd.x + self.v * rd.y;
