+++ src/colorize.rs
@@ -3,30 +3,23 @@ use crate::{
     hitable::Hitable,
     pdf::{HitablePDF, MixturePDF},
     ray::Ray,
+    scenes::Scene,
     Float, SHADOW_EPSILON,
 };
 use rand::prelude::*;
 use std::sync::Arc;
 
 /// The main coloring function
-pub fn colorize(
-    ray: &Ray,
-    background_color: Color,
-    world: &Hitable,
-    lights: Arc<Hitable>, // NOTE: possibly hitablelist, or bvhnode, or something new?
-    depth: u32,
-    max_depth: u32,
-    rng: ThreadRng,
-) -> Color {
+pub fn colorize(ray: &Ray, scene: &Scene, depth: u32, max_depth: u32, rng: ThreadRng) -> Color {
     if depth > max_depth {
         // Ray bounce limit reached, return background_color
-        return background_color;
+        return scene.background_color;
     }
 
     // Here, smoothing is used to avoid "shadow acne"
-    match world.hit(&ray, SHADOW_EPSILON, Float::MAX, rng) {
+    match scene.objects.hit(&ray, SHADOW_EPSILON, Float::MAX, rng) {
         // If the ray hits nothing, return the background color.
-        None => background_color,
+        None => scene.background_color,
 
         // Hit something
         Some(hit_record) => {
@@ -50,9 +43,7 @@ pub fn colorize(
                             scatter_record.attenuation
                                 * colorize(
                                     &scatter_record.specular_ray.unwrap(), // should always have a ray at this point
-                                    background_color,
-                                    world,
-                                    lights,
+                                    scene,
                                     depth + 1,
                                     max_depth,
                                     rng,
@@ -60,8 +51,10 @@ pub fn colorize(
                         }
                         crate::materials::MaterialType::Diffuse => {
                             // Use a probability density function to figure out where to scatter a new ray
-                            let light_ptr =
-                                HitablePDF::new(Arc::clone(&lights), hit_record.position);
+                            let light_ptr = HitablePDF::new(
+                                Arc::clone(&scene.priority_objects), // TODO: fix silly arc
+                                hit_record.position,
+                            );
                             let mixture_pdf = MixturePDF::new(light_ptr, scatter_record.pdf_ptr);
 
                             let scattered =
@@ -69,15 +62,7 @@ pub fn colorize(
                             let pdf_val = mixture_pdf.value(scattered.direction, ray.time, rng);
 
                             // recurse
-                            let recurse = colorize(
-                                &scattered,
-                                background_color,
-                                world,
-                                lights,
-                                depth + 1,
-                                max_depth,
-                                rng,
-                            );
+                            let recurse = colorize(&scattered, scene, depth + 1, max_depth, rng);
 
                             // Blend it all together
                             emitted
